Queue module reused and addapted
The queue module was borrowed from the T1 from the same course
(INFO1019), because in order to implement the Optimum algorithm,
there needed to be a queue, not a list (as previously implemented).
But there were some key adaptations for the use in this project.
The module may become more generic (making use of C preprocessors
perhaps?) in the future for such adaptations not being necessary
anymore.

Adaptations:
* queue node holds unsigned long, not void
* queue node create function accepts 0 as information
* queue node getinfo returns unsigned long through pointer
* queue node getinfo returns queue ret condition
* queue module test adapted to the previous changes
* queue module test inherits test module interface