Remove queue module + WIP Optimal algorithm
I found that queues would be too expensive (both on memory
and on perfomance), so I scratched a new algorithm that might
be even better. It's O(l) - being l the number of lines in
the file - just like the other one I was thinking about, in
terms of memory complexity. But in reality, it's half of the
memory. To access a node from the queue, it's O(n). For an
array it's O(1). A lot of improvement.